feat: Window Move

+++ Assets/01.Script/WinAPI/WindowHandle.cs
@@ -3,12 +3,12 @@ using System.Runtime.InteropServices;
 
 public static class WindowHandle
 {
-    private static IntPtr windowHandle;
+    private static int windowHandle;
     
     [DllImport("user32.dll")]
-    private static extern IntPtr GetActiveWindow();
+    private static extern int GetActiveWindow();
     
-    public static IntPtr GetWindowHandle()
+    public static int GetWindowHandle()
     {
         if (windowHandle == null)
         {
